+++ Wrappers/Python/cil/plugins/astra/operators/__init__.py
@@ -3,4 +3,4 @@ from .AstraProjector3DSimple import AstraProjector3DSimple
 from .AstraProjector3DMC import AstraProjector3DMC
 from .AstraProjectorMC import AstraProjectorMC
 from .AstraProjectorFlexible import AstraProjectorFlexible
-from .AstraOperator import AstraOperator
+from .ProjectionOperator import ProjectionOperator
