+++ Wrappers/Python/test/test_ProjectorConsistency.py
@@ -20,7 +20,6 @@ from cil.framework import ImageGeometry, AcquisitionGeometry
 from cil.framework import ImageData, AcquisitionData
 
 from cil.plugins.astra.operators import AstraProjectorSimple, AstraProjector3DSimple, AstraProjectorFlexible
-from cil.plugins.astra.operators import AstraOperator
 from cil.utilities.display import plotter2D
